Add Hardware Offload Validation as a test option for flows.
Hardware Offload Validation uses "perf3", "tcpdump" and "ethtool"
to determine whether a flow has been hardware offloaded properly.
The expectation is that little to no packets are received on the
VF representors for the client and/or server pods.

In order to get the VF representor for the pods involved, the ft-tools
pod is used with the "crictl" application to get the POD ID.

The ft-tools pod was updated to include "crictl". Also the tools
daemonset was updated to mount the host filesystem (similar to
openshift's debug pod e.g. "oc debug node/<name>"). Also fixed the
node selector to point to "sriov", otherwise the ft-tools won't be
created for the nodes.

Some flows aren't expected to be hardware offloaded; thus the "HWOL"
option would not perform the validation for those flows.